Interface gráfica atualizada.

+++ CG-SGI/include/geometria/Ponto.h
@@ -8,19 +8,19 @@ class Ponto : public ObjetoGeometrico {
 public:
 	Ponto();
 	Ponto(const Ponto& ponto);
-	Ponto(const String& nome, const int x, const int y, const int z);
+	Ponto(const String& nome, const double x, const double y, const double z);
 	virtual ~Ponto();
 	void operator=(const Ponto& ponto);
-	int getX() const;
-	int getY() const;
-	int getZ() const;
+	double getX() const;
+	double getY() const;
+	double getZ() const;
 	QList<Ponto> getPontos();
 	const String toString() const;
 
 protected:
-	int coord_x;
-	int coord_y;
-	int coord_z;
+	double coord_x;
+	double coord_y;
+	double coord_z;
 
 };
